+++ app/Http/Controllers/NoteController.php
@@ -14,8 +14,13 @@ use Inertia\Response;
 
 class NoteController extends Controller
 {
+    /**
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function index(Project $project): Response
     {
+        $this->authorize('view', $project);
+
         $meta = [
             'title' => 'Notes',
         ];
@@ -24,23 +29,40 @@ class NoteController extends Controller
         return Inertia::render('Project/ShowNotes', compact('project', 'meta'));
     }
 
+    /**
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function store(Project $project, NoteRequest $request): RedirectResponse
     {
+        $this->authorize('manage', $project);
+
         $note = new Note($request->validated());
         $note->project()->associate($project);
         $note->save();
 
         return back()->with('success');
     }
+
+    /**
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function update(Project $project, Note $note, NoteRequest $request): RedirectResponse
     {
+        $this->authorize('manage', $project);
+
         $note->fill($request->validated());
         $note->save();
 
         return back()->with('success');
     }
+
+    /**
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function destroy(Project $project, Note $note): RedirectResponse
     {
+        $this->authorize('manage', $project);
+
         $note->delete();
 
         return back()->with('success');
